+++ flask/api/booking/booking.py
@@ -1,12 +1,8 @@
 from flask import Blueprint,render_template,request,jsonify
 from api import db,bcrypt,jwt
 from api.models.bookings_model import Booking
-from api.models.attractions_model import Attraction
 from api.models.members_model import Member
-from flask_jwt_extended import (
-    JWTManager, jwt_required, create_access_token,
-    create_refresh_token, get_jwt_identity,unset_access_cookies,unset_refresh_cookies,set_access_cookies,set_refresh_cookies
-)
+from flask_jwt_extended import (jwt_required, get_jwt_identity)
 
 booking = Blueprint("booking",
     __name__,
@@ -21,37 +17,11 @@ def frontPage():
 @jwt_required(fresh=True)
 def getBookingData():
     try:
-        booking_lists = []
-        identity = get_jwt_identity()
-        member_email =identity
-        query_booking_lists = db.session.query(Booking, Attraction,Member). \
-            select_from(Booking).join(Attraction).join(Member).filter(Booking.member_email == member_email,Booking.order_number == None).all()
-        if len(query_booking_lists) == 0:
-            query = Member.query.filter_by(email=member_email).first()
-            return jsonify(status="noData",name=query.name)
-        for booking, attraction, member in query_booking_lists:
-            image_urls = []
-            for image in attraction.images:
-                image_urls.append(image.image_url)
-            booking_data = {
-                "data": {
-                    "member":{
-                        "name": member.name,
-                        "email": member.email
-                    },
-                    "attraction": {
-                        "id": attraction.id,
-                        "name": attraction.name,
-                        "address": attraction.address,
-                        "image": image_urls
-                    },
-                    "id": booking.id,
-                    "date": booking.date,
-                    "time": booking.time,
-                    "price": booking.price
-                }
-            }
-            booking_lists.append(booking_data)
+        member_email = get_jwt_identity()
+        booking_lists = Member.get_not_pay_booking_data(member_email)
+        if booking_lists == "no_data":
+            query = Member.get_member_auth_data(member_email)
+            return jsonify(status="noData",name=query["data"]["name"])
         return jsonify(booking_lists)
     except Exception as ex:
         return jsonify(error="true",message=f"{ex}"),500
@@ -61,22 +31,12 @@ def getBookingData():
 def sendBookingData():
     try:
         data = request.get_json()
-        attraction_id = data["attractionId"]
-        date = data["date"]
-        time = data["time"]
-        price = data["price"]
-        identity = get_jwt_identity()
-        member_email =identity
-        print("Booking Data:",data,member_email)
-        # 存到資料庫
-        filters = {"member_email": member_email, "date": date, "time": time}
-        result = Booking.query.filter_by(**filters).all()
+        member_email = get_jwt_identity()
+        result = Booking.check_booking_data_by_time(data, member_email)
         if len(result) >=1:
             return jsonify(error="true",message="⚠ 已在這個時段預約行程"),400
-        print(result)
-        data = Booking(member_email, attraction_id, date, time, price)
-        db.session.add(data)
-        db.session.commit()
+
+        Booking.build_new_booking_data(data, member_email)
         return jsonify(ok="true"),200
     except Exception as ex:
         return jsonify(error="true",message=f"{ex}"),500
@@ -87,10 +47,7 @@ def deleteBookingData():
     try:
         data = request.get_json()
         member_email = get_jwt_identity()
-        booking_id = data["bookingId"]
-        query = Booking.query.filter_by(member_email=member_email,id=booking_id).first()
-        db.session.delete(query)
-        db.session.commit()
+        query = Booking.delete_booking_data_by_bookingId(data, member_email)
         return jsonify(ok="true")
     except Exception as ex:
         return jsonify(error="true",message=f"{ex}"),500
