+++ flask/api/models/orders_model.py
@@ -11,7 +11,6 @@ class Orders(db.Model):
     contact_phone = db.Column(db.String(30), nullable=False)
     price = db.Column(db.Integer, nullable=False)
 
-
     def __init__(self, order_number,member_email, contact_name, contact_email, contact_phone, price):
         self.order_number  = order_number
         self.member_email = member_email
@@ -19,3 +18,8 @@ class Orders(db.Model):
         self.contact_email = contact_email
         self.contact_phone = contact_phone
         self.price = price
+    
+    def build_new_order_data(number, member_email, contact_name, contact_email, contact_phone, order_price):
+        order_upload_data = Orders(number, member_email, contact_name, contact_email, contact_phone, order_price)
+        db.session.add(order_upload_data)
+        db.session.commit()
