+++ flask/api/orders/orders.py
@@ -1,10 +1,9 @@
-from flask import Flask,Blueprint,jsonify,request,render_template
-from api import db,bcrypt,jwt
+from flask import Blueprint,jsonify,request
+from api import db
 from api.models.bookings_model import Booking
 from api.models.attractions_model import Attraction
-from api.models.members_model import Member
 from api.models.orders_model import Orders
-from flask_jwt_extended import (JWTManager, jwt_required,get_jwt_identity)
+from flask_jwt_extended import (jwt_required,get_jwt_identity)
 import time
 import re
 import os 
@@ -23,8 +22,7 @@ orders = Blueprint("orders",
 @jwt_required(fresh=True)
 def create_new_order():
     try:
-        identity = get_jwt_identity()
-        member_email =identity
+        member_email = get_jwt_identity()
         data = request.get_json()
         contact_email = data["order"]["contact"]["email"]
         contact_phone = data["order"]["contact"]["phone"]
@@ -38,8 +36,7 @@ def create_new_order():
         for i in range(data_length):
             booking_id = data["order"]["trip"][i]["attraction"]["bookingId"]
             booking_id_list.append(booking_id)
-            # 要用join處理
-            query_order_number = Booking.query.filter(Booking.id == booking_id_list[i], Booking.order_number != None).all()
+            query_order_number = Booking.get_payed_bookingId(booking_id_list[i])
             result = len(query_order_number)
             if result != 0:
                 return jsonify(error="true",message="⚠ 請勿重複付款"),400
@@ -75,12 +72,9 @@ def create_new_order():
                     }
                 }
             }
-            order_upload_data = Orders(number, member_email, contact_name, contact_email, contact_phone, order_price)
-            db.session.add(order_upload_data)
-            db.session.commit()
+            Orders.build_new_order_data(number, member_email, contact_name, contact_email, contact_phone, order_price)
             for i in range(data_length):
-                booking_upload_data = Booking.query.filter_by(id=booking_id_list[i]).update({"order_number":number})
-                db.session.commit()
+                Booking.update_booking_order_number(booking_id_list[i], number)
             return success,200
         else:
             return jsonify(error="true",message=req.json()["msg"])
@@ -93,41 +87,9 @@ def create_new_order():
 @jwt_required(fresh=True)
 def get_new_order(orderNumber):
     try:
-        trip_lists = []
-        query_booking_lists = db.session.query(Orders,Booking, Attraction). \
-            select_from(Orders).join(Booking).join(Attraction).filter(Orders.order_number == orderNumber).all()
-        if len(query_booking_lists) == 0:
+        orders_query_data  = Booking.get_order_data_by_id(orderNumber)
+        if orders_query_data  == "no_data":
             return jsonify(error="true",message="沒有此訂單編號"),400
-        for orders, booking ,attraction in query_booking_lists:
-            image_urls = []
-            for image in attraction.images:
-                image_urls.append(image.image_url)
-            orders_query_data = { 
-                "data": {
-                    "number": orders.order_number,
-                    "price": orders.price,
-                    "trip":trip_lists,
-                    "contact": {
-                        "name": orders.contact_name,
-                        "email": orders.contact_email,
-                        "phone": orders.contact_phone
-                    },
-                    "status": 1
-                }
-            }
-            trip = {
-                "attraction": {
-                    "id": attraction.id,
-                    "name": attraction.name,
-                    "address": attraction.address,
-                    "image": image_urls
-                },
-                "booking_id": booking.id,
-                "booking_price": booking.price,
-                "date": booking.date,
-                "time": booking.time
-            }
-            trip_lists.append(trip)
         return orders_query_data   
     except Exception as ex:
         return jsonify(error="true",message=f"{ex}"),500
